requirements.txt
numpy
pandas
scikit-learn
scipy
umap-learn
